clean code + wait kuzzle start before asking for status

+++ scripts/run-tests.sh
@@ -1,25 +1,39 @@
 #!/bin/bash
-set -e
 
-SCRIPT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
+#-------------------------------------------------------------------------------
+#
+#   Kuzzle end-to-end test sandbox
+#
+#   Script aim: run end-to-end test
+#   - start chaos mode if needed
+#   - run proxy end-to-end tests if exists
+#   - run kuzzle core end-to-end tests if exists
+#   - run backoffice end-to-end
+#
+#-------------------------------------------------------------------------------
+
+set -e
 
 COLOR_END="\e[39m"
 COLOR_BLUE="\e[34m"
 COLOR_YELLOW="\e[33m"
 
+SANDBOX_DIR="/tmp/sandbox"
+SCRIPT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
+
 echo -e "[$(date --rfc-3339 seconds)] - ${COLOR_BLUE}Starting functional testing suite...$COLOR_END"
 
-if [[ $ENABLE_CHAOS_MODE == "true" ]]; then
-  bash "$SCRIPT_DIR/run-chaos.sh" &
+if [[ "${ENABLE_CHAOS_MODE}" == "true" ]]; then
+  bash "${SCRIPT_DIR}/parts/run-chaos.sh" &
 fi
 
-pushd "/tmp/sandbox" &>/dev/null
+pushd "${SANDBOX_DIR}" &>/dev/null
   # run e2e proxy tests if exists
   pushd kuzzle-proxy &>/dev/null
     echo -e "[$(date --rfc-3339 seconds)] - ${COLOR_BLUE}Running kuzzle-proxy tests...$COLOR_END"
 
     if [[ $(npm run | grep functional-testing) ]]; then
-      bash -c "${SCRIPT_DIR}/reset.sh"
+      bash -c "${SCRIPT_DIR}/parts/reset-kuzzle.sh"
 
       echo -e "[$(date --rfc-3339 seconds)] - ${COLOR_BLUE}Running kuzzle-proxy functional tests...$COLOR_END"
 
@@ -37,7 +51,7 @@ pushd "/tmp/sandbox" &>/dev/null
     echo -e "[$(date --rfc-3339 seconds)] - ${COLOR_BLUE}Running kuzzle tests...$COLOR_END"
 
     if [[ $(npm run | grep functional-testing) ]]; then
-      bash -c "${SCRIPT_DIR}/reset.sh"
+      bash -c "${SCRIPT_DIR}/parts/reset-kuzzle.sh"
 
       echo -e "[$(date --rfc-3339 seconds)] - ${COLOR_BLUE}Running kuzzle functional tests...$COLOR_END"
 
@@ -51,29 +65,32 @@ pushd "/tmp/sandbox" &>/dev/null
 
   # run e2e kuzzle-backoffice tests with chrome & firefox
   pushd kuzzle-backoffice &>/dev/null
-    bash -c "${SCRIPT_DIR}/reset.sh"
+    bash -c "${SCRIPT_DIR}/parts/reset-kuzzle.sh"
 
-    echo -e "[$(date --rfc-3339 seconds)] - ${COLOR_BLUE}Running kuzzle backoffice tests (chrome)...$COLOR_END"
+    if [[ "${TESTIM_PROJECT}" == "" ]] || [[ "${TESTIM_TOKEN}" == "" ]]; then
+      echo -e "[$(date --rfc-3339 seconds)] - ${COLOR_YELLOW}Skipping backoffice tests, you need to define TESTIM_PROJECT and TESTIM_TOKEN environment variables.$COLOR_END"
+    else
+      echo -e "[$(date --rfc-3339 seconds)] - ${COLOR_BLUE}Running kuzzle backoffice tests (chrome)...$COLOR_END"
 
-    docker inspect "testim" &>/dev/null && sh -c "docker kill testim || true" && sh -c "docker rm -vf testim || true"
+      docker inspect "testim" &>/dev/null && sh -c "docker kill testim || true" && sh -c "docker rm -vf testim || true"
 
-    docker run --network="bridge" \
-      --name "testim" \
-      --link "hub:hub" \
-      --link "proxy:proxy" \
-      --link "backoffice:backoffice" \
-      --volume "/tmp/sandbox/kuzzle-backoffice/test/e2e/run-test.sh:/opt/run-test.sh" \
-      --volume "/tmp/sandbox/kuzzle-backoffice/test/e2e/config-file.js:/opt/config-file.js" \
-      -e "TESTIM_PROJECT=$TESTIM_PROJECT" \
-      -e "TESTIM_TOKEN=$TESTIM_TOKEN" \
-      -e "BROWSER=chrome" \
-      -ti \
-      tests/kuzzle-base \
-        bash -c 'chmod 755 /opt/run-test.sh && /opt/run-test.sh'
+      docker run --network="bridge" \
+        --name "testim" \
+        --link "hub:hub" \
+        --link "proxy:proxy" \
+        --link "backoffice:backoffice" \
+        --volume "${SANDBOX_DIR}/kuzzle-backoffice/test/e2e/run-test.sh:/opt/run-test.sh" \
+        --volume "${SANDBOX_DIR}/kuzzle-backoffice/test/e2e/config-file.js:/opt/config-file.js" \
+        -e "TESTIM_PROJECT=${TESTIM_PROJECT}" \
+        -e "TESTIM_TOKEN=${TESTIM_TOKEN}" \
+        -e "BROWSER=chrome" \
+        -ti \
+        tests/kuzzle-base \
+          bash -c 'chmod 755 /opt/run-test.sh && /opt/run-test.sh'
 
       echo -e "[$(date --rfc-3339 seconds)] - ${COLOR_BLUE}Kuzzle backoffice tests ended (chrome)...$COLOR_END"
 
-      bash -c "${SCRIPT_DIR}/reset.sh"
+      bash -c "${SCRIPT_DIR}/parts/reset.sh"
 
       echo -e "[$(date --rfc-3339 seconds)] - ${COLOR_BLUE}Running kuzzle backoffice tests (firefox)...$COLOR_END"
 
@@ -84,16 +101,16 @@ pushd "/tmp/sandbox" &>/dev/null
         --link "hub:hub" \
         --link "proxy:proxy" \
         --link "backoffice:backoffice" \
-        --volume "/tmp/sandbox/kuzzle-backoffice/test/e2e/run-test.sh:/opt/run-test.sh" \
-        --volume "/tmp/sandbox/kuzzle-backoffice/test/e2e/config-file.js:/opt/config-file.js" \
-        -e "TESTIM_PROJECT=$TESTIM_PROJECT" \
-        -e "TESTIM_TOKEN=$TESTIM_TOKEN" \
+        --volume "${SANDBOX_DIR}/kuzzle-backoffice/test/e2e/run-test.sh:/opt/run-test.sh" \
+        --volume "${SANDBOX_DIR}/kuzzle-backoffice/test/e2e/config-file.js:/opt/config-file.js" \
+        -e "TESTIM_PROJECT=${TESTIM_PROJECT}" \
+        -e "TESTIM_TOKEN=${TESTIM_TOKEN}" \
         -e "BROWSER=firefox" \
         -ti \
         tests/kuzzle-base \
           bash -c 'chmod 755 /opt/run-test.sh && /opt/run-test.sh'
 
       echo -e "[$(date --rfc-3339 seconds)] - ${COLOR_BLUE}Kuzzle backoffice tests ended (firefox)...$COLOR_END"
-
+    fi
   popd &>/dev/null
 popd &>/dev/null
